requirements.txt
django==4.1.2
djangorestframework==3.14.0
django-filter==22.1
django-nose==1.4.7
coverage==6.5.0
defusedxml==0.7.1
pylint==2.15.4
